+++ server/data/migrations/20190115141457_create_subscriptions_table.js
@@ -2,8 +2,8 @@ exports.up = function(knex, Promise) {
     return knex.schema.createTable("subscriptions", table => {
         table.increments();
         table.string("transaction_id").notNullable();
-        table.integer("date_created").notNullable();
-        table.integer("duration").defaultTo(2592000000);
+        table.bigInteger("date_created").notNullable();
+        table.bigInteger("duration").defaultTo(2592000000);
         table.integer("user_id").notNullable().unsigned().references("id").inTable("users");
     });
 };
